+++ config/domainscoza.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'url' => env('DOMAINSCOZA_URL'),
+    'username' => env('DOMAINSCOZA_USERNAME'),
+    'password' => env('DOMAINSCOZA_PASSWORD'),
+];