+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `domainscoza-api` will be documented in this file.
+
+## 2022-10-24
+
+- First version
+- Can login and retrieve list of domains
+- Test suite
+- Laravel Facade